pixie_server

Function find_mac

pub(crate) fn find_mac(ip: Ipv4Addr) -> Result<MacAddr6>

Finds the mac address for the given ip.

This function searches the address in the arp cache, if it is not available it tries to populate it by pinging the peer.