pixie_server

Module state

The database for pixie-server.

Structure of the storage directory:

config.yaml: configuration file for pixie-server

registered.json: json file containing all information about registered units.

admin/: directory containing the static files for the admin web interface.

chunks/: directory containing the image’s chunks.

images/: directory containing the image’s info.

tftpboot/: directory containing the necessary files for network boot.

Re-exports

pub use units::UnitSelector;

Modules

images 🔒

units 🔒

Structs

State

See the module-level documentation.

Constants

CHUNKS_DIR 🔒

CONFIG_YAML 🔒

IMAGES_DIR 🔒

REGISTERED_JSON 🔒

Functions

atomic_write 🔒

Atomically write data at the specified path.

build_hostmap 🔒

Builds a map from ip address to hostname parsing the hostfile at path.